backend, frontend
switching to vite
backend - finished completely
frontend -
maintenance finished
tank layout finished
fish catalog finished
tank calculations formatted
ai response presentation formated

+++ backend/models/aqualayout_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, JSON, ForeignKey
+from sqlalchemy import Column, Integer, String, DateTime, JSON, ForeignKey, Float
 from sqlalchemy.sql import func
 from backend.db.base import Base
 from pydantic import BaseModel, ConfigDict, EmailStr
@@ -14,9 +14,9 @@ class AquaLayout(Base):
     owner_email = Column(String, ForeignKey('users.email'), nullable=False)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     tank_name = Column(String, nullable=False)
-    tank_length = Column(Integer, nullable=False)  # in inches
-    tank_width = Column(Integer, nullable=False)
-    tank_height = Column(Integer, nullable=False)
+    tank_length = Column(Float, nullable=False)  # in cm (stored as decimal)
+    tank_width = Column(Float, nullable=False)  # in cm (stored as decimal)
+    tank_height = Column(Float, nullable=False)  # in cm (stored as decimal)
     water_type = Column(String, nullable=False)  # Freshwater or Saltwater
     fish_data = Column(JSON, nullable=False)  # [{"name": "Goldfish", "quantity": 2}]
     comments = Column(String, nullable=True)
@@ -31,9 +31,9 @@ class FishEntry(BaseModel):
 class AquaLayoutCreate(BaseModel):
     owner_email: EmailStr
     tank_name: str
-    tank_length: int
-    tank_width: int
-    tank_height: int
+    tank_length: float
+    tank_width: float
+    tank_height: float
     water_type: str
     fish_data: List[FishEntry]
     comments: Optional[str] = None
@@ -44,9 +44,9 @@ class AquaLayoutResponse(BaseModel):
     owner_email: EmailStr
     created_at: datetime
     tank_name: str
-    tank_length: int
-    tank_width: int
-    tank_height: int
+    tank_length: float
+    tank_width: float
+    tank_height: float
     water_type: str
     fish_data: List[FishEntry]
     comments: Optional[str] = None
